requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
redis==5.0.8
motor==3.7.1
python-dotenv==1.0.1
uvloop==0.19.0; platform_system != 'Windows'
pytest==8.2.1
networkx==3.2.1
pyvis==0.3.2